+++ pymc3/distributions/special.py
@@ -3,7 +3,7 @@ import theano.tensor as tt
 from theano.scalar.basic_scipy import GammaLn, Psi, I0, I1
 from theano import scalar
 
-__all__ = ['gammaln', 'multigammaln', 'psi', 'i0', 'i1', 'log_i0']
+__all__ = ['gammaln', 'multigammaln', 'psi', 'log_i0']
 
 scalar_gammaln = GammaLn(scalar.upgrade_to_float, name='scalar_gammaln')
 gammaln = tt.Elemwise(scalar_gammaln, name='gammaln')
@@ -12,14 +12,16 @@ gammaln = tt.Elemwise(scalar_gammaln, name='gammaln')
 def multigammaln(a, p):
     """Multivariate Log Gamma
 
-    :Parameters:
-        a : tensor like
-        p : int degrees of freedom
-            p > 0
+    Parameters
+    ----------
+    a : tensor like
+    p : int
+       degrees of freedom. p > 0
     """
     i = tt.arange(1, p + 1)
     return (p * (p - 1) * tt.log(np.pi) / 4.
             + tt.sum(gammaln(a + (1. - i) / 2.), axis=0))
+
 
 def log_i0(x):
     """
@@ -32,11 +34,6 @@ def log_i0(x):
                                   + 9. / (128. * x**2.) + 225. / (3072. * x**3.)
                                   + 11025. / (98304. * x**4.)))
 
+
 scalar_psi = Psi(scalar.upgrade_to_float, name='scalar_psi')
 psi = tt.Elemwise(scalar_psi, name='psi')
-
-scalar_i0 = I0(scalar.upgrade_to_float, name='scalar_i0')
-i0 = tt.Elemwise(scalar_i0, name='i0')
-
-scalar_i1 = I1(scalar.upgrade_to_float, name='scalar_i1')
-i1 = tt.Elemwise(scalar_i1, name='i1')